requirements.txt
Django
djangorestframework
PyJWT
django-cors-headers
djangorestframework-simplejwt
